+++ History.md
@@ -4,12 +4,12 @@
 
 ### Breaking changes
 
-Success and error handler options on LastFmUpdate and LastFmInfo have been removed and replaced with more generic `handlers` option.
+Event handler options on RecentTrackStream, LastFmUpdate and LastFmInfo have been removed and replaced with more generic `handlers` option.
 
 ### New features
 
 * More accurate error reporting from RecentTrackParser
-* Added request() method
+* Added request() method. Provides low-level support for entire Last.Fm API.
 
 ## 0.5.0